I've upgraded the chain to allow resolvers in a bit more seamless way, so you can use your chain in any order, as long as in the end there's a resolver.
It's still possible to use the resolver chain function, but be aware that now it expects a function as the second argument.
Also a resolver must be the last element, else you'll try and add elements to a resolved chain which in turn will be gracefully ignored.
Just so you know.

This behavior should be improved in the future, but it's good enough for a POC. (for now)
I'm still playing around with it :)